Update, start integrating with react
Make revisions use a specific RevId type.
Add revisions to DataSource, for use in renderers to detect changes to data.
Rename MapStateSTM state monad to MapState (from S) for clarity.
Move ContextDemo to DataContext in tree-react. and adapt to using a DataSource. This works, but we need to adapt to using ViewOps instead of manual reporting of Ids used.
Move MapDataSource from SpecUtils to its own non-test class, for use in demo of DataContext.