Add js project to handle extraction of muiapi.json
This uses react-docgen api for more control, and uses the same logic in the material-ui API build to filter files and post-processes the generated json. This handles the new chainPropType and deprecation warnings, and also picks up inheritance via the annotation. This should give a cleaner result, avoids producing a facade for internal components, and can be updated in future to reflect any changes in material-ui's own processing of react-docgen output. Next step is to check that the generated json is parsed properly by code generation, and add support for the inheritance information and new "componentPropType" custom PropType.