Use inheritance data from muiapi.json
This allows us to remove the old (and incomplete) inheritance map in context, and use the actual inheritance from annotations in material-ui.

This adds quite a lot of missing props.

The code for inheritance is now also quite a lot neater, using List[(String, Prop)] for all props data, and combining similar functions into one.

This also introduces preprocessing of the component map, to allow us to apply a small number of tweaks to inheritance and naming of components where needed.