Provide separate context for transactor
Rather than just a DataContext, we now use ReactDataContexts that has a Context[ReactData] and a Context[Transactor].
This allows ViewPC access to a Transactor to be provided implicitly to Cursor, enabling transactions with the Cursor (ViewC already had access via ReactData). This can then be used in cases where a View only needs data from the Cursor itself, not from other Ids in the data. Added an example of this to edit the text in TodoItems in LocalIndexedDataRootDemo.