Move ClientState back to tree-ot, and complete.
Add some methods to Operation, and update the docs.
Split Operation.after into two forms - one to apply to operations transformed to be applied on the server (which take priority) and one for operations from the client. This gives priority to whichever operation reaches the server first, in the case of ties.
Extend and correct the notes in comments in ServerState.
Add OTSpec test for tie-breaking behaviour of Operation.transform.
Add first test of client and server state working together.