Remove priority from Operation, add only where needed
PriorityOperation is an operation with a priority - used to provide remote operations to client state, and store history in server state.
OpRev now has priority as well as operation and revision.
Reinstate MapDataSource for use in tests.
ClientState now stores priority
Fix Diff - remove priority, and retain the common end to generate an Operation with correct length.