Neaten up RevId, stability
RevId now contains only the TransactionId, since it's not necessary to add a transaction clock - it only matters which transaction most recently updated a value, not the details of when in that transaction it was updated.
Remove unused MapDataSource.
This also fixes issue with MapStateSTM flagging of unstable transactions - before any call to createGuid was considered to be a U op, even the ones used to create guids for RevIds. Now we don't need to do any U op, just get the transaction id from context.
Next step is most likely to just use TransactionId instead of RevId - it's not important to have the phantom A type, we don't actually need to keep track of the type of data here.