Optimise DataComponent and LocalDataRoot
LocalDataRoot does not need the full ref/effect system for making state available to ReactTransactor - we can just use the "transform" version of setState instead, and setState is stable so does not need to be stored in a ref.

DataComponent should useMemo rather than useRef, to avoid creating and discarding a new inner component on each render (since useRef is not lazy).

Update README with description of tree-slinky and todo covering optimisation checks.