Reinstate cursor updates
Make getting a cursor update count as viewing the guid of the OTList, so that views are updated on any cursor update. It's probably not a common use case, but a view might theoretically just display a cursor position, rather than the actual OTList, and so we need to make sure we update in this case.